+++ lib/td_dq/rules/rule_result.ex
@@ -9,14 +9,15 @@ defmodule TdDq.Rules.RuleResult do
     field :date, :utc_datetime
     field :result, :integer
     field :parent_domains, :string, default: ""
-
+    field :errors, :integer
+    field :records, :integer
     timestamps()
   end
 
   @doc false
   def changeset(%RuleResult{} = rule_result, attrs) do
     rule_result
-    |> cast(attrs, [:implementation_key, :date, :parent_domains, :result])
+    |> cast(attrs, [:implementation_key, :date, :parent_domains, :result, :errors, :records])
     |> validate_required([:implementation_key, :date, :result])
   end
 end